+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 ## master
 
+1.2.0-22
+
+### Bugfixes
+- Fix melody-stream not rerendering in some cases [#112](https://github.com/trivago/melody/pull/112)
+
+### Chore & Maintenance
+- Removes `Chai` and `Sinon` support, Migrates tests to use `Jest`'s matchers. [#103](https://github.com/trivago/melody/pull/103)
+
+## 1.2.0-5 (beta)
+
+### Features
+
+- Introduce `melody-streams` API [#102](https://github.com/trivago/melody/pull/102)
+
 ## 1.2.0-4 (beta)
 
 ### Features
@@ -11,7 +25,6 @@
 - Introduce `useAtom` hook [#79](https://github.com/trivago/melody/pulls/79)
 - Introduce `useStore` hook and performance marks for hooks [#98](https://github.com/trivago/melody/pulls/98)
 - Added async mounting of components [#82](https://github.com/trivago/melody/pull/82)
-- Introduce `melody-streams` API [#102](https://github.com/trivago/melody/pull/102)
 
 ### Fixes